% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{Cum_Hfx}
\alias{Cum_Hfx}
\title{Cumulative hazard}
\usage{
Cum_Hfx(SURVIVAL, t)
}
\arguments{
\item{SURVIVAL}{a SURVIVAL object}

\item{t}{Time}
}
\value{
a cumulative hazard at time t
}
\description{
Cumulative hazard at time t using the cumulative hazard method
from the SURVIVAL object
}
