% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_class.R
\name{EXPONENTIAL_hazard}
\alias{EXPONENTIAL_hazard}
\alias{EXPONENTIAL_surviving}
\alias{EXPONENTIAL_failure}
\title{Parameters for a exponential distribution}
\usage{
EXPONENTIAL_hazard(lambda)

EXPONENTIAL_surviving(surv, t)

EXPONENTIAL_failure(fail, t)
}
\arguments{
\item{lambda}{hazard parameter}

\item{surv}{proportion surviving at time t}

\item{t}{time t}

\item{fail}{proportion fail at time t}
}
\value{
an object of class SURVIVALPARAM, PEXPONENTIAL
}
\description{
Create an object with parameters for a survival exponential distribution
}
\section{Functions}{
\itemize{
\item \code{EXPONENTIAL_surviving}: Survival proportion at time t

\item \code{EXPONENTIAL_failure}: Failure proportion at time t
}}

