% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{ggplot_random}
\alias{ggplot_random}
\title{Plot random draws from the distribution}
\usage{
ggplot_random(SURVIVAL, timeto, subjects, nsim, alpha = 0.1)
}
\arguments{
\item{SURVIVAL}{a SURVIVAL object}

\item{timeto}{plot the distribution up to timeto}

\item{subjects}{number of subjects to simulate in each simulation}

\item{nsim}{number of simulations}

\item{alpha}{alpha value for the graph}
}
\description{
a ggPlot with random survival times using the random generator in the
SURVIVAL object
}
