% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_class.R
\name{GOMPERTZ_hazard}
\alias{GOMPERTZ_hazard}
\alias{GOMPERTZ_surviving}
\alias{GOMPERTZ_failure}
\title{Parameters for a Gompertz distribution}
\usage{
GOMPERTZ_hazard(scale, shape)

GOMPERTZ_surviving(surv, t, shape)

GOMPERTZ_failure(fail, t, shape)
}
\arguments{
\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{surv}{proportion surviving at time t}

\item{t}{time t}

\item{fail}{proportion fail at time t}
}
\value{
an object
}
\description{
Create an object with parameters for a Gompertz distribution
}
\section{Functions}{
\itemize{
\item \code{GOMPERTZ_surviving}: Survival at time t

\item \code{GOMPERTZ_failure}: Failure at time t
}}

