% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{invCum_Hfx}
\alias{invCum_Hfx}
\title{Inverse of the cumulative hazard}
\usage{
invCum_Hfx(H, SURVIVAL)
}
\arguments{
\item{H}{cumulative hazard}

\item{SURVIVAL}{a SURVIVAL object}
}
\description{
Return the time t for a give cumulative hazard using the method
from the SURVIVAL object
}
