% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linsum}
\alias{linsum}
\title{Linear sum within a data.frame}
\usage{
linsum(data, vars, coeffs)
}
\arguments{
\item{data}{the data.frame with the variables}

\item{vars}{the variables  to select from the data.frame}

\item{coeffs}{the coefficients to apply}
}
\value{
a vector with the sum of each variable value multiply by the coefficient.
}
\description{
Perform a sum with the linear combination between the variables
and the coefficients
}
