% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_exponential_class.R
\name{PIECEWISE_hazards}
\alias{PIECEWISE_hazards}
\alias{PIECEWISE_surviving}
\alias{PIECEWISE_failure}
\title{Parameters for a piece wise exponential distribution}
\usage{
PIECEWISE_hazards(breaks, hazards)

PIECEWISE_surviving(surv, breaks, segments)

PIECEWISE_failure(fail, breaks, segments)
}
\arguments{
\item{breaks}{breaks that indicates a change in hazard}

\item{hazards}{hazards values valid up to the next break}

\item{surv}{proportion surviving at last break point}

\item{segments}{hazard scale valid up to the next break}

\item{fail}{proportion fail at last break point}
}
\value{
an object of class SURVIVALPARAM, PPIECEWISE_EXP
}
\description{
Create an object with parameters for a survival piece wise exponential distribution
}
\section{Functions}{
\itemize{
\item \code{PIECEWISE_surviving}: Survival at last segment

\item \code{PIECEWISE_failure}: Failure at last segment
}}

