% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot a SURVIVAL OBJECT}
\usage{
plot_survival(SURVIVAL, timeto)
}
\arguments{
\item{SURVIVAL}{a SURVIVAL object}

\item{timeto}{timeto used in the graphs}
}
\description{
Plot of the functions in the survival object until timeto t
}
