% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{rsurv}
\alias{rsurv}
\title{Random survival times}
\usage{
rsurv(n, SURVIVAL)
}
\arguments{
\item{n}{number of observations}

\item{SURVIVAL}{a SURVIVAL object}
}
\value{
random survival times
}
\description{
Return random survival times following the distribution defined in the SURVIVAL
object
}
