% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{rsurvdf}
\alias{rsurvdf}
\title{Random survival times for a hazard rate}
\usage{
rsurvdf(SURVIVAL, hr)
}
\arguments{
\item{SURVIVAL}{a SURVIVAL object}

\item{hr}{a vector with hazard rates.}
}
\value{
random survival times
}
\description{
Return random survival times following the distribution defined in the SURVIVAL
object, and assuming a proportionality of the hazards, the log(time) is
modified by the hazard ratio. A hazard ratio of 1 means not change to the
baseline hazard of the distribution. A hazard ratio lower than 1 means longer
times and hazard ratios higher than 1 means shorter times.
frame
}
