% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{sfx}
\alias{sfx}
\title{Survival proportion}
\usage{
sfx(t, SURVIVAL)
}
\arguments{
\item{t}{Time}

\item{SURVIVAL}{a SURVIVAL object}
}
\value{
Proportion surviving at time t
}
\description{
Survival proportion at time t using the method from the SURVIVAL object
}
