% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{sfx}
\alias{sfx}
\alias{hfx}
\alias{Cum_Hfx}
\alias{invCum_Hfx}
\alias{rsurv}
\alias{rsurvhr}
\alias{rsurvaft}
\alias{rsurveh}
\alias{plot_survival}
\alias{ggplot_survival_random}
\alias{compare_survival}
\alias{ggplot_survival_hr}
\alias{ggplot_survival_aft}
\alias{ggplot_survival_eh}
\title{Functions for SURVIVAL objects}
\usage{
sfx(SURVIVAL, t)

hfx(SURVIVAL, t)

Cum_Hfx(SURVIVAL, t)

invCum_Hfx(SURVIVAL, H)

rsurv(SURVIVAL, n)

rsurvhr(SURVIVAL, hr)

rsurvaft(SURVIVAL, aft)

rsurveh(SURVIVAL, aft, hr)

plot_survival(SURVIVAL, timeto, main)

ggplot_survival_random(SURVIVAL, timeto, subjects, nsim, alpha = 0.1)

compare_survival(SURVIVAL1, SURVIVAL2, timeto, main)

ggplot_survival_hr(SURVIVAL, hr, timeto, subjects, nsim, alpha = 0.1)

ggplot_survival_aft(SURVIVAL, aft, timeto, subjects, nsim, alpha = 0.1)

ggplot_survival_eh(SURVIVAL, aft, hr, timeto, subjects, nsim, alpha = 0.1)
}
\arguments{
\item{SURVIVAL}{a SURVIVAL object}

\item{t}{Time}

\item{H}{cumulative hazard}

\item{n}{number of observations}

\item{hr}{hazard ratio}

\item{aft}{accelerated failure time}

\item{timeto}{plot the distribution up to timeto}

\item{main}{title of the graph}

\item{subjects}{number of subjects per group to simulate in each simulation}

\item{nsim}{number of simulations}

\item{alpha}{alpha value for the graph}

\item{SURVIVAL1}{a SURVIVAL object}

\item{SURVIVAL2}{a SURVIVAL object}
}
\value{
Depending on the function a proportion surviving, hazard,
cumulative hazard, inverse of the cumulative hazard, a random draw or a plot
}
\description{
All the SURVIVAL objects have access to the functions described here
}
\details{
\code{rsurveh()} generates random values under the Extended Hazards model of
Chen & Jewell (2001), which combines a proportional-hazards effect and an
accelerated-failure-time effect: h*(t) = hr * aft * h0(aft * t). It nests
the proportional hazards model (\code{aft = 1}) and the accelerated failure
time model (\code{hr = 1}) as special cases. Note this differs from the
Accelerated Hazards model of Chen & Wang (2000), h*(t) = h0(theta * t),
which is the special case \code{hr = 1 / aft}.
}
\section{Functions}{
\itemize{
\item \code{sfx()}: Survival function

\item \code{hfx()}: Hazard function

\item \code{Cum_Hfx()}: Cumulative Hazard function

\item \code{invCum_Hfx()}: Inverse of the Cumulative Hazard function

\item \code{rsurv()}: Generate random values from the distribution

\item \code{rsurvhr()}: Generate random values from the distribution under proportional hazard ratios

\item \code{rsurvaft()}: Generate random values from the distribution under accelerated failure time ratios

\item \code{rsurveh()}: Generate random values from the distribution under the Extended Hazards model (combined proportional hazards and accelerated failure time)

\item \code{plot_survival()}: Plot of the survival functions

\item \code{ggplot_survival_random()}: ggplot of the simulation of survival times

\item \code{compare_survival()}: Compare graphically two survival distributions

\item \code{ggplot_survival_hr()}: ggplot of the simulation of survival times with hazard ratios

\item \code{ggplot_survival_aft()}: ggplot of the simulation of survival times with accelerated failure times

\item \code{ggplot_survival_eh()}: ggplot of the simulation of survival times with the Extended Hazards model

}}
\examples{
#' # Define a SURVIVAL object
obj <- s_factory(s_weibull, surv = 0.8, t = 2, shape = 1.2)

# Survival, Hazard and Cumulative hazard at time 0.4
sfx(SURVIVAL = obj, t= 0.4)
hfx(SURVIVAL = obj, t = 0.4)
Cum_Hfx(SURVIVAL = obj, t = 0.4)

# Time when the Cumulative hazard is 0.8
invCum_Hfx(SURVIVAL = obj, H = 0.8)

# Draw one random survival time from the distribution
rsurv(SURVIVAL = obj, n = 1)

# Draw one random survival time from the distribution under Proportional
# hazard, Accelerated failure time or the combined Extended Hazards model.
rsurvhr(SURVIVAL = obj, hr = 0.5)
rsurvaft(SURVIVAL = obj, aft = 2)
rsurveh(SURVIVAL = obj, aft = 2, hr = 0.5)

# Plot the survival functions
plot_survival(SURVIVAL = obj, timeto = 2, main = "Example of Weibull distribution" )

}
