% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_class.R
\name{WEIBULL_hazard}
\alias{WEIBULL_hazard}
\alias{WEIBULL_surviving}
\alias{WEIBULL_failure}
\alias{WEIBULL_surverg}
\title{Parameters for a Weibull distribution}
\usage{
WEIBULL_hazard(scale, shape)

WEIBULL_surviving(surv, t, shape)

WEIBULL_failure(fail, t, shape)

WEIBULL_surverg(logscale, logshape)
}
\arguments{
\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{surv}{proportion surviving at time t}

\item{t}{time t}

\item{fail}{proportion fail at time t}

\item{logscale}{scale parameter from a survreg model}

\item{logshape}{shape parameter from a survreg model}
}
\value{
an object of class SURVIVALPARAM, PWEIBULL
}
\description{
Create an object with parameters for a Weibull exponential distribution
}
\section{Functions}{
\itemize{
\item \code{WEIBULL_surviving}: Survival at time t

\item \code{WEIBULL_failure}: Failure at time t

\item \code{WEIBULL_surverg}: From survreg model
}}

