% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{hfx}
\alias{hfx}
\title{Hazard}
\usage{
hfx(t, SURVIVAL)
}
\arguments{
\item{t}{Time}

\item{SURVIVAL}{a Survival object}
}
\value{
the hazard at time t
}
\description{
Hazard a time t using the hazard method from the SURVIVAL object
}
