% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_single_number}
\alias{is_single_number}
\title{Confirm is a single number}
\usage{
is_single_number(x)
}
\arguments{
\item{x}{a variable to evaluate}
}
\value{
\code{TRUE} if is a single number, \code{FALSE} otherwise
}
\description{
Evaluates if the argument is a single number
}
\examples{
is_single_number(3)  #TRUE
is_single_number(c(3,3,3)) #FALSE
is_single_number(list(a=3)) #FALSE
is_single_number("3") #FALSE
}
