% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{rsurvdf}
\alias{rsurvdf}
\title{Random survival times for a data frame}
\usage{
rsurvdf(.data, vars, coeffs, SURVIVAL)
}
\arguments{
\item{.data}{a data.frame}

\item{vars}{a character vector with names of variables in the data.frame}

\item{coeffs}{a numeric vector with the value of the coefficient to multiply
to the corresponding variable}

\item{SURVIVAL}{a SURVIVAL object}
}
\value{
random survival times
}
\description{
Return random survival times following the distribution defined in the SURVIVAL
object, and assuming a proportionality of the hazards, the log(time) is
modified by the linear combination of the vars and coeffs values in the data
frame
}
