% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_class.R
\name{s_factory}
\alias{s_factory}
\title{SURVIVAL class object's factory generic}
\usage{
s_factory(s_family, ...)
}
\arguments{
\item{s_family}{an object of class S_FAMILY}

\item{...}{parameters to define the survival distribution}
}
\value{
a SURVIVAL object
}
\description{
Creates a SURVIVAL object using the information from the SURVIVALPARAM parameter
This is implemented by each subclass of the SURVIVALPARAM, i.e.
PEXPONENTIAL, PWEIBULL, PGOMPERTZ PPIECEWISE.
}
\note{
The created object has the following methods (functions)

sfx(t): S function that returns the survival proportion at a given time

hfx(t): A function that returns the hazard at a given time

Cum_Hfx(t): A function that returns the cumulative hazard at a given time

invHFx(H): A function the returns the time given a cumulative hazard value

rsurv(x): A function that returns random times from the survival distribution

rsurvhr(hr): A function that returns random times from the survival distribution
but following a proportional hazard assumption. For example a rsurvhr(c(1,0.5))
will produce a random survival time in which the first observation has a hazard
ratio of 1 and the second a hazard ratio of 0.5.
}
