% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_class.R,
%   R/piecewise_exponential_class.R, R/survival_class.R, R/weibull_class.R
\name{SURVIVAL_factory.PGOMPERTZ}
\alias{SURVIVAL_factory.PGOMPERTZ}
\alias{SURVIVAL_factory.PPIECEWISE_EXP}
\alias{SURVIVAL_factory}
\alias{SURVIVAL_factory.default}
\alias{SURVIVAL_factory.PWEIBULL}
\title{SURVIVAL class object's factory generic}
\usage{
\method{SURVIVAL_factory}{PGOMPERTZ}(PSURVIVAL, ...)

\method{SURVIVAL_factory}{PPIECEWISE_EXP}(PSURVIVAL, ...)

SURVIVAL_factory(PSURVIVAL, ...)

\method{SURVIVAL_factory}{default}(PSURVIVAL, ...)

\method{SURVIVAL_factory}{PWEIBULL}(PSURVIVAL, ...)
}
\arguments{
\item{PSURVIVAL}{an object of class SURVIVALPARAMS}

\item{...}{Currently not used}
}
\value{
a SURVIVAL object
}
\description{
Creates a SURVIVAL object using the information from the SURVIVALPARAM parameter
This is implemented by each subclass of the SURVIVALPARAMS, ie
for PEXPONENTIAL, PWEIBULL, PPIECEWISEEXP.
}
\section{Methods (by class)}{
\itemize{
\item \code{PGOMPERTZ}: Gompertz distribution

\item \code{PPIECEWISE_EXP}: Piece wise exponential distribution

\item \code{default}: Default implementation

\item \code{PWEIBULL}: Weibull distribution
}}

\note{
The created object has the following methods (functions)

sfx(t): S function that returns the survival proportion at a given time

hfx(t): A function that returns the hazard at a given time

Cum_Hfx(t): A function that returns the cumulative hazard at a given time

invHFx(H): A function the returns the time given a cumulative hazard value

rsurv(x): A function that returns random times from the survival distribution

rsurvdf(.data, vars, coeffs): A function that returns random times from the survival distribution
but following a proportional hazard assumption, the log(time) is modified by the
linear combination of the parameters given by the variables specified in vars and
the coefficients specified in coeffs.
}
