% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survobj-package.R
\docType{package}
\name{survobj-package}
\alias{survobj}
\alias{survobj-package}
\title{survobj: Objects to simulate survival times}
\description{
Functions to generate objects that simulate survival times.
}
\author{
\strong{Maintainer}: Aponte John \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID})

}
\keyword{internal}
