% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survobj-package.R
\docType{package}
\name{survobj-package}
\alias{survobj}
\alias{survobj-package}
\title{survobj: Objects to Simulate Survival Times}
\description{
Generate objects that simulate survival times. Random values for the distributions are generated using the method described by Bender (2003) \url{https://epub.ub.uni-muenchen.de/id/eprint/1716} and Leemis (1987) in Operations Research, 35(6), 892–894.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://johnaponte.github.io/survobj/}
  \item \url{https://github.com/johnaponte/survobj}
  \item Report bugs at \url{https://github.com/johnaponte/survobj/issues}
}

}
\author{
\strong{Maintainer}: Aponte John \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID}) [copyright holder]

}
\keyword{internal}
